Portable int/long conversion across Python 2 / Python 3
Differential Revision: https://reviews.llvm.org/D59585

llvm-svn: 356671

+++ lldb/examples/python/jump.py
@@ -78,7 +78,7 @@ def parse_linespec(linespec, frame, result):
         if (mo is not None):
             matched = True
             # print "Matched <address-expression>"
-            address = long(mo.group(1), base=0)
+            address = int(mo.group(1), base=0)
             breakpoint = target.BreakpointCreateByAddress(address)
 
     if (not matched):
